requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.2
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.7
django-cors-headers==4.7.0
django-debug-toolbar==5.0.1
django-filter==25.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-taggit==6.1.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
gunicorn==20.1.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
kavenegar==1.1.2
kombu==5.5.0
MarkupSafe==3.0.2
openapi-codec==1.3.2
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.50
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
setuptools==77.0.3
simplejson==3.20.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
whitenoise==6.9.0
